Restructured sanity.config.ts for easier readability
