+++ cms/structure/index.ts
@@ -1,0 +1,46 @@
+import {StructureBuilder} from 'sanity/structure'
+import {ArrowDownIcon, CalendarIcon, DocumentTextIcon, HomeIcon, UserIcon} from '@sanity/icons'
+
+const SINGLETONS = [{id: 'frontpage', title: 'Forside', _type: 'frontpage'}]
+
+const LANGUAGES = [
+  {id: `nb`, title: `🇳🇴 Norwegian (Bokmål)`},
+  {id: `en`, title: `🇬🇧 English`},
+]
+
+export const deskStructure = (S: StructureBuilder) =>
+  S.list()
+    .title('Innhold')
+    .items([
+      ...SINGLETONS.map((singleton) =>
+        S.listItem()
+          .title(singleton.title)
+          .id(singleton.id)
+          .icon(HomeIcon)
+          .child(
+            S.list()
+              .title(singleton.title)
+              .id(singleton.id)
+
+              .items(
+                LANGUAGES.map((language) =>
+                  S.documentListItem()
+                    .schemaType(`frontpage`)
+                    .id(`${singleton.id}-${language.id}`)
+                    .title(`${singleton.title} (${language.id.toLocaleUpperCase()})`),
+                ),
+              )
+              .canHandleIntent(
+                (intentName, params) => intentName === 'edit' && params.id.startsWith(singleton.id),
+              ),
+          ),
+      ),
+      S.documentTypeListItem('article').title('Artikler').icon(DocumentTextIcon),
+      S.documentTypeListItem('event').title('Forestillinger').icon(CalendarIcon),
+      S.documentTypeListItem('role').title('Roller').icon(UserIcon),
+      S.listItem()
+        .title('Footer')
+        .id('footer')
+        .child(S.document().schemaType('footer').documentId('footer'))
+        .icon(ArrowDownIcon),
+    ])
